--- simulators/portal-bridge-tests/src/test-data/README.md
@@ -1,2 +1,0 @@
-### Test Data Collection of Forks Blocks
-This test includes the headers, bodies, receipts for blocks 1, 100, 7000000, 15600000 (post-merge), and 17510000 (post-shanghai).
