+++ simulators/trin-bridge/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "portal-bridge-tests"
+name = "trin-bridge"
 version = "0.1.0"
 authors = ["https://github.com/ethereum/portal-hive/graphs/contributors"]
 edition = "2021"
@@ -10,7 +10,7 @@ portal-bridge = { git = "https://github.com/ethereum/trin" }
 trin-validation = { git = "https://github.com/ethereum/trin" }
 hivesim = { path = "../../hivesim-rs" }
 itertools = "0.10.5"
-serde_json = "1.0.87"
+serde_yaml = "0.9"
 tokio = { version = "1", features = ["full"] }
 tracing = "0.1.37"
 tracing-subscriber = "0.3.16"
