+++ simulators/trin-bridge/Dockerfile
@@ -1,15 +1,15 @@
 FROM rust:1.71.1 AS builder
 
 # create a new empty shell project
-RUN USER=root cargo new --bin portal-bridge-tests
-WORKDIR /portal-bridge-tests
+RUN USER=root cargo new --bin trin-bridge
+WORKDIR /trin-bridge
 
 RUN apt-get update && apt-get install clang -y
 
 # copy over manifests and source to build image
-COPY ./simulators/portal-bridge-tests/Cargo.lock ./Cargo.lock
-COPY ./simulators/portal-bridge-tests/Cargo.toml ./Cargo.toml
-COPY ./simulators/portal-bridge-tests/src ./src
+COPY ./simulators/trin-bridge/Cargo.lock ./Cargo.lock
+COPY ./simulators/trin-bridge/Cargo.toml ./Cargo.toml
+COPY ./simulators/trin-bridge/src ./src
 COPY ./hivesim-rs ./../../hivesim-rs
 
 # build for release
@@ -28,15 +28,15 @@ RUN dpkg -i libssl1.1_1.1.0g-2ubuntu4_amd64.deb
 #RUN dpkg -i libssl1.1_1.1.1f-1ubuntu2_arm64.deb
 
 # copy build artifacts from build stage
-COPY --from=builder /portal-bridge-tests/target/release/portal-bridge-tests .
-COPY --from=builder /portal-bridge-tests/src/test-data ./test-data
+COPY --from=builder /trin-bridge/target/release/trin-bridge .
+COPY --from=builder /trin-bridge/src/test-data ./test-data
 
 RUN apt-get update && apt-get install libcurl4 -y
 
 ENV RUST_LOG=debug
 
-# add fake secrets for bridge activation
+# Fake secrets for Trin bridge activation, not actually used in the used `BridgeMode::Test` mode
 ENV PANDAOPS_CLIENT_ID=xxx
 ENV PANDAOPS_CLIENT_SECRET=xxx
 
-ENTRYPOINT ["./portal-bridge-tests"]
+ENTRYPOINT ["./trin-bridge"]
